feat(PaymentHistoryRestController)
- 결제 이력의 경우 결제와 다른 Aggregate이기 때문에 따로 분리
- 외부로부터의 접근은 오직 Aggregate Root를 통하게만 만들어 응집도를 높이기 위함
- 서비스 레이어 분리를 통해 향후 결제에서 이용될 트랜잭션 제어를 위함
